Prevent “directory already exists error”
* Not gets the "directory already exists error" over and over

+++ makefile
@@ -1,5 +1,5 @@
 all:
-	mkdir ./bin
+	mkdir -p ./bin
 	gcc -std=c11 src/*.c -I./include -o bin/student
 clean:
-	rm ./bin -rf
+	if [ -d "./bin" ]; then rm ./bin -rf; fi
